NON BACKWARD COMPATIBLE: Changed syntax of 'Suites'-key in .repos files
This is done as preparation for more enhanced .repos features coming later.
The main advantiage of the new way of describing suites is, that that
the same suitenames could be listed several times now (makes only sense
if additional metadata differs) which was not possible with the 'suitename
as key' syntax as before.