modernized example config main.repos and disabled gpg-signature verification
to make things easier. If you want gpg-signature verification, you need
to ensure that the required gpg keyring can be found somewhere on your
system (I can't resolve this dependency for you in a distribution
independent way).

Please also note that newer versions of libapt somtimes deny to accept old
signatures even if the keyring is correctly provided (depends on the
length of the Hash used for the signature).